moved input elements to folder
