+++ app/assets/javascripts/components/inputs/checkboxes.jsx
