+++ app/assets/javascripts/components/manifest/manifest_element_form.jsx
@@ -1,6 +1,6 @@
 import React from 'react'
-import InputField from '../form_inputs/input_field'
-import TextField from '../form_inputs/text_field'
+import InputField from '../inputs/input_field'
+import TextField from '../inputs/text_field'
 import { Result } from './manifest_results'
 
 // Edit component for a single variable/expression from a manifest
