+++ app/assets/javascripts/components/plots/forms/heatmap.jsx
@@ -1,7 +1,7 @@
 import React, { Component } from 'react';
 import PlotForm, { subscribePlotInputField, commonfields, matrixConsignmentKeyFilter } from './plot_form';
-import Select from '../../form_inputs/select';
-import InputField from '../../form_inputs/input_field';
+import Select from '../../inputs/select';
+import InputField from '../../inputs/input_field';
 
 const defaultHeatmapPlot = {
   name: '',
